Fixed field form validation for whitelist host.
