+++ src/main/java/org/jenkinsci/plugins/gwt/whitelist/WhitelistVerifier.java
@@ -8,14 +8,12 @@ import com.github.jgonian.ipmath.Ipv6;
 import com.github.jgonian.ipmath.Ipv6Range;
 import com.google.common.base.Optional;
 import com.google.common.net.InetAddresses;
-import hudson.util.FormValidation;
 import java.util.List;
 import java.util.Map;
 import org.jenkinsci.plugins.gwt.global.CredentialsHelper;
 import org.jenkinsci.plugins.gwt.global.Whitelist;
 import org.jenkinsci.plugins.gwt.global.WhitelistItem;
 import org.jenkinsci.plugins.plaincredentials.StringCredentials;
-import org.kohsuke.stapler.QueryParameter;
 
 public class WhitelistVerifier {
 
@@ -165,57 +163,5 @@ public class WhitelistVerifier {
     }
     throw new WhitelistException(
         "Sending host \"" + remoteHost + "\" was not matched by whitelist.");
-  }
-
-  private Boolean validateIpValue(String ipValue) {
-    Boolean isValid = false;
-
-    Boolean isCIDR = false;
-    Boolean isRange = false;
-
-    String[] hostParts = ipValue.split("/");
-
-    if (hostParts.length == 2) {
-      isCIDR = true;
-    } else {
-      hostParts = ipValue.split("-");
-      if (hostParts.length == 2) {
-        isRange = true;
-      }
-    }
-
-    if (isCIDR || isRange) {
-      int leftValueLength = InetAddresses.forString(hostParts[0]).getAddress().length;
-      if (leftValueLength == 4) {
-        if (Ipv4Range.parse(ipValue) != null) {
-          isValid = true;
-        }
-      } else if (leftValueLength == 16) {
-        if (Ipv6Range.parse(ipValue) != null) {
-          isValid = true;
-        }
-      }
-    } else {
-      int ipValueLength = InetAddresses.forString(hostParts[0]).getAddress().length;
-      if (ipValueLength == 4) {
-        if (Ipv4.parse(ipValue) != null) {
-          isValid = true;
-        }
-      } else if (ipValueLength == 16) {
-        if (Ipv6.parse(ipValue) != null) {
-          isValid = true;
-        }
-      }
-    }
-
-    return isValid;
-  }
-
-  public FormValidation checkIP(@QueryParameter String value) {
-    if (validateIpValue(value)) {
-      return FormValidation.ok();
-    } else {
-      return FormValidation.error("IP Address must be in IPV4 or IPV6 CIDR or IP range notation.");
-    }
   }
 }
